File split router
