+++ src/router/callback.route.ts
@@ -1,0 +1,9 @@
+import { RouteRecordRaw } from "vue-router";
+
+export const CallbackRoute = [
+	{
+		path: "/oauth2",
+		name: "OAuth2Callback",
+		component: () => import("@views/OAuth2.vue"),
+	},
+] as RouteRecordRaw[];
