+++ src/router/home.route.ts
@@ -1,0 +1,8 @@
+import { RouteRecordRaw } from "vue-router";
+
+export const HomeRoute = {
+	path: "/",
+	name: "Home",
+	meta: { transparentNav: true },
+	component: () => import("@views/Home/Home.vue"),
+} as RouteRecordRaw;
