+++ src/router/users.route.ts
@@ -1,0 +1,32 @@
+import { useActorStore } from "@/store/actor";
+import { Permissions } from "@/structures/Role";
+import { User } from "@/structures/User";
+import { RouteRecordRaw } from "vue-router";
+
+export const UsersRoute = [
+	{
+		path: "/users/:userID",
+		name: "User",
+		props: true,
+		component: () => import("@views/UserPage/UserPage.vue"),
+		children: [
+			{
+				path: "settings",
+				name: "UserSettings",
+				component: () => import("@views/UserSettings/UserSettings.vue"),
+				props: true,
+				beforeEnter: (to, _, next) => {
+					const actor = useActorStore();
+					if (
+						User.HasPermission(actor.user, Permissions.ManageUsers) ||
+						to.params.userID === actor.user?.id
+					) {
+						next();
+					} else {
+						next(`/users/${to.params.userID}`);
+					}
+				},
+			},
+		],
+	},
+] as RouteRecordRaw[];
